refactor(github service): :truck: move Github REST API client to a dedicated package
